rename MachineUtil to IciwiUtil
