+++ .project
@@ -22,7 +22,7 @@
 	</natures>
 	<filteredResources>
 		<filter>
-			<id>1673510517516</id>
+			<id>0</id>
 			<name></name>
 			<type>30</type>
 			<matcher>
