+++ build.gradle
@@ -8,7 +8,7 @@ version '1.19.3a1'
 compileJava.options.encoding 'UTF-8'
 
 repositories {
-  maven {url 'https://papermc.io/repo/repository/maven-public/'}
+  maven { url 'https://papermc.io/repo/repository/maven-public/' }
   maven { url 'https://ci.mg-dev.eu/plugin/repository/everything' }
   maven { url 'https://jitpack.io' }
   mavenLocal()
