+++ src/main/java/mikeshafter/iciwi/api/IciwiPlugin.java
@@ -1,0 +1,13 @@
+package mikeshafter.iciwi.api;
+
+import mikeshafter.iciwi.util.IciwiCard;
+import org.bukkit.plugin.Plugin;
+
+public interface IciwiPlugin extends Plugin {
+  /**
+   * Gets the plugin's fare card class
+   *
+   * @return Class of the fare card
+   */
+  Class<? extends IcCard> getFareCardClass();
+}
