+++ src/main/java/mikeshafter/iciwi/tickets/TicketMachine.java
@@ -18,7 +18,7 @@ import java.time.Instant;
 import java.util.*;
 import java.util.function.Function;
 
-import static mikeshafter.iciwi.util.MachineUtil.*;
+import static mikeshafter.iciwi.util.IciwiUtil.*;
 
 public class TicketMachine implements Machine {
