+++ src/main/java/mikeshafter/iciwi/util/IciwiUtil.java
@@ -13,7 +13,7 @@ import java.util.List;
 import java.util.regex.Pattern;
 
 
-public class MachineUtil {
+public class IciwiUtil {
   public static boolean isDouble(String s) {
     return Pattern.matches(("[\\x00-\\x20]*"+"[+-]?("+"NaN|"+"Infinity|"+"((((\\d+)(\\.)?((\\d+)?)([eE][+-]?(\\d+))?)|"+"(\\.(\\d+)([eE][+-]?(\\d+))?)|"+"(("+"(0[xX](\\p{XDigit}+)(\\.)?)|"+"(0[xX](\\p{XDigit}+)?(\\.)(\\p{XDigit}+))"+")[pP][+-]?(\\d+)))"+"[fFdD]?))"+"[\\x00-\\x20]*"), s);
   }
